fix several build issues
