+++ .github/workflows/ci.yml
@@ -118,11 +118,16 @@ jobs:
       - name: Install dependencies
         run: npm install
 
+      - name: Build dependencies
+        run: |
+          cd packages/config-loader && npm run build || true
+          cd ../cloud && npm run build || true
+          
       - name: Build and test scripts
         run: |
           cd packages/scripts
           npm run build
-          npx tsc --project tsconfig.check.json --noEmit
+          npx tsc --noEmit --skipLibCheck
 
   test-comprehensive:
     name: Comprehensive Test Suite
