+++ packages/cloud/tsconfig.json
@@ -32,7 +32,8 @@
     "isolatedModules": true,
     "typeRoots": [
       "./node_modules/@types"
-    ]
+    ],
+    "types": ["node"]
   },
   "exclude": [
     "node_modules",
