--- packages/scripts/tsconfig.check.json
@@ -1,18 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "noEmit": true,
-    "jsx": "react"
-  },
-  "include": [
-    "*.ts",
-    "lib/**/*.ts",
-    "lib/**/*.tsx",
-    "utils/**/*.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "dist",
-    "../cloud/**/*"
-  ]
-}
